+++ server/db/models/database.js
@@ -1,5 +1,7 @@
 'use strict';
-var Sequelize = require('sequelize')
+var Sequelize = require('sequelize');
+var pg = require('pg');
+
 
 module.exports = function (db) {
 
@@ -18,6 +20,17 @@ module.exports = function (db) {
             },
             URI: function() {
                 return 'postgres://localhost:5432/' + this.dbName;
+            }
+        },
+        classMethods: {
+            makeClientDatabase: function(createdDB) {
+                var dbName = createdDB.dbName;
+                var connectionString = 'postgres://localhost:5432/masterDB';
+                var client = new pg.Client(connectionString);
+                client.connect();
+                var query = client.query('CREATE DATABASE ' + dbName);
+                query.on('end', function() { client.end(); });
+                return createdDB; 
             }
         }
     });
